make into extension(kinda), and other stuff
