+++ public/scripts/extensions/variables/manifest.json
@@ -1,0 +1,11 @@
+{
+    "display_name": "Variables",
+    "loading_order": 99,
+    "requires": [],
+    "optional": [],
+    "js": "index.js",
+    "css": "",
+    "author": "IkariDev",
+    "version": "1.0.0",
+    "homePage": "https://github.com/SillyTavern/SillyTavern"
+}
